requirements_enhanced.txt
streamlit==1.29.0
pandas==2.1.4
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2

# For spaCy English model (install separately):
# python -m spacy download en_core_web_sm